refactor(frontend): extract issue details sheet
